Move function apis
